[add] property success

+++ src/main/java/com/realEstate/realEstate/repository/PropertyRepository.java
@@ -12,6 +12,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PropertyRepository extends JpaRepository<Property, Integer> {
+public interface PropertyRepository extends JpaRepository<Property, Long> {
     public Page<Property> findAllByUser(User user, Pageable pageable);
 }
